requirements.txt
discord.py
jishaku
aiosqlite